change tests_models -> test_model, modified test_view, add test_urls
